+++ src/pool_thread/message_loop.rs
@@ -62,10 +62,10 @@ where
                                 new_pool_item.name()
                             );
                             self.pool_item_hash_map.insert(id, new_pool_item);
-                            AddResponse::new(id, true, None)
+                            AddResponse::new(id, Ok(()))
                         }
                         Err(new_pool_item_error) => {
-                            AddResponse::new(id, false, Some(new_pool_item_error.error_message))
+                            AddResponse::new(id, Err(new_pool_item_error.error_message))
                         }
                     }
                     .into()