+++ src/request_response/id_targeted.rs
@@ -29,9 +29,8 @@ mod tests {
     #[test]
     #[should_panic(expected = "not expected; only requests are required to support IdTargeted")]
     fn request_response_contains_response_request_panics() {
-        let target = RequestResponse::<Randoms, RandomsAddRequest>::Response(AddResponse::new(
-            0, true, None,
-        ));
+        let target =
+            RequestResponse::<Randoms, RandomsAddRequest>::Response(AddResponse::new(0, Ok(())));
 
         target.id();
     }