+++ tests/example_simple.rs
@@ -16,7 +16,7 @@ pub fn example_simple_one_level_thread_pool() {
     thread_pool
         .send_and_receive((0..1000usize).map(RandomsAddRequest))
         .expect("thread pool to be available")
-        .for_each(|response: AddResponse| assert!(response.success()));
+        .for_each(|response: AddResponse| assert!(response.result().is_ok()));
 
     // now create 1000 messages asking them for the sum of the Randoms objects contained random numbers
     // The message will be routed to the thread to where the targeted object resides
@@ -41,13 +41,13 @@ pub fn example_simple_one_level_thread_pool() {
     thread_pool
         .send_and_receive(iter::once(RemovePoolItemRequest(1)))
         .expect("thread pool to be available")
-        .for_each(|response: RemovePoolItemResponse| assert!(response.success()));
+        .for_each(|response: RemovePoolItemResponse| assert!(response.item_existed()));
 
     // add a new pool item with id 1000
     thread_pool
         .send_and_receive(iter::once(RandomsAddRequest(1000)))
         .expect("thread pool to be available")
-        .for_each(|response: AddResponse| assert!(response.success()));
+        .for_each(|response: AddResponse| assert!(response.result().is_ok()));
 
     // all pool items are dropped when the basic thread pool batcher is dropped
     // the threads are shutdown and joined back the the main thread