+++ CHANGELOG.md
@@ -33,4 +33,6 @@ The PoolItem trait provides a cleaner and hopefully more intuitive interface wit
 - Change send to return a Result and propagate result where necessary
 - Make error_message in NewPoolItemError public
 - Add SendAndReceiverRawMock to deal with heterogeneous streams of messages
+- Change success() to item_existed() in RemovePoolItemResponse
+- Change success() to result in AddResponse